WIP,  built with python 3.11, removed pybullet and habitat dependencies.

+++ polymetis/src/clients/test_grpc_client.cpp
@@ -115,7 +115,8 @@ private:
     ClientContext context;
     Status status = stub_->ControlUpdate(&context, state_, &torque_command);
     if (!status.ok()) {
-      spdlog::error("SendCommand failed: {}", status.error_code());
+      spdlog::error("SendCommand failed: {}",
+                    static_cast<int>(status.error_code()));
       return false;
     }
     return true;
